Refactor: Move base classes to common module
- Move src/agent/agents/base.py to src/agent/common/base.py
- Create src/agent/common module for shared infrastructure
- Update all imports to use new location
- Improve conceptual organization: common/ for infrastructure, agents/ for implementations
- No functional changes, purely organizational improvement
