+++ src/agent/common/__init__.py
@@ -1,0 +1,13 @@
+"""
+Common utilities and base classes for the agent system.
+
+This module contains foundational infrastructure shared across all agents,
+including base classes, common types, and utility functions.
+"""
+
+from src.agent.common.base import Agent, AgentResult
+
+__all__ = [
+    "Agent",
+    "AgentResult",
+]
