+++ src/agent/common/base.py
